Share copyFile helper across Unix and Windows updaters
Move the duplicated file-copy logic into update/copy.go so both
platform installers stay in sync during maintenance.
